requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
asgiref==3.4.1; python_version >= '3.6'
backports.zoneinfo==0.2.1; python_version < '3.9'
certifi==2021.10.8
charset-normalizer==2.0.9; python_version >= '3'
django-cors-headers==3.10.1
django==4.0
djangorestframework==3.13.0
idna==3.3; python_version >= '3'
pendulum==2.1.2
psycopg2-binary==2.9.2
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2021.3
pytzdata==2020.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pyyaml==6.0
requests==2.26.0
sentry-sdk==1.5.1
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.4.2; python_version >= '3.5'
urllib3==1.26.7; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
